backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "backend"
version = "0.1.0"
description = ""
authors = ["flaviaserret04 <flaviaserret04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.6"
djangorestframework = "^3.16.1"
drf-spectacular = "^0.28.0"
drf-nested-routers = "^0.95.0"
packaging = "^25.0"
platformdirs = "^4.4.0"
djangorestframework-simplejwt = "^5.5.1"
tomlkit = "^0.13.3"
django-cors-headers = "^4.9.0"
django-health-check = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
]
